Adjusted tests to reviewed production code
