+++ src/graphql/schema/uuid/abstract-uuid/utils.ts
@@ -19,15 +19,11 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { ApolloError } from 'apollo-server'
 import * as R from 'ramda'
 
-import { SerloDataSource } from '../../../data-sources/serlo'
-import { resolveConnection } from '../../connection'
-import { isDefined } from '../../utils'
 import { EntityRevisionType, EntityType } from '../abstract-entity'
 import { AbstractUuidPayload, UuidResolvers } from '../abstract-uuid'
-import { CommentPayload, ThreadData, ThreadDataType } from '../thread'
+import { createThreadResolvers } from '../thread/utils'
 import { DiscriminatorType } from './types'
 
 const validTypes = [
@@ -41,40 +37,5 @@ export function isUnsupportedUuid(payload: AbstractUuidPayload) {
 }
 
 export function createUuidResolvers(): UuidResolvers {
-  return {
-    async threads(parent, cursorPayload, { dataSources }) {
-      const { firstCommentIds } = await Promise.resolve(
-        dataSources.serlo.getThreadIds({ id: parent.id })
-      )
-
-      return resolveConnection({
-        nodes: await Promise.all(
-          firstCommentIds.map((id) => toThreadPayload(dataSources.serlo, id))
-        ),
-        payload: cursorPayload,
-        createCursor(node) {
-          return node.commentPayloads[0].id.toString()
-        },
-      })
-    },
-  }
-}
-
-export async function toThreadPayload(
-  serlo: SerloDataSource,
-  firstCommentId: number
-): Promise<ThreadData> {
-  const firstComment = await serlo.getUuid<CommentPayload>({
-    id: firstCommentId,
-  })
-  if (firstComment === null) {
-    throw new ApolloError('There are no comments yet')
-  }
-  const remainingComments = await Promise.all(
-    firstComment.childrenIds.map((id) => serlo.getUuid<CommentPayload>({ id }))
-  )
-  return {
-    __typename: DiscriminatorType.Thread,
-    commentPayloads: [firstComment, ...remainingComments.filter(isDefined)],
-  }
+  return createThreadResolvers()
 }
